refactor: move toast logic to custom hook
cleanup

clean up

Delete yarn.lock

add yarn.lock

revert yarn.lock file

revert
