+++ src/services/version.ts
@@ -1,1 +1,1 @@
-export const LIB_VERSION = '1.5.0';
+export const LIB_VERSION = 'git';
